Break out of `get_group_info()` if rate limit exceeded
This took me way too long because Google's documentation for this
library is crap. These are the official docs for the `HttpError` class:
https://googleapis.github.io/google-api-python-client/docs/epy/googleapiclient.errors.HttpError-class.html

Absolutely no details on the different fields provided by the class
(which I need to search for "exceeded" in the error response). Had to
look into the actual implementation:
https://github.com/googleapis/google-api-python-client/blob/d5cf4e0b1537eeafdb96ae831bbdffc5f1c6a94c/googleapiclient/errors.py#L35

(also, HttpError.reason is generated using a different method to the the
"reason" in _should_retry_response() function, which is run as part of
API calls and logs warnings:
https://github.com/googleapis/google-api-python-client/blob/cbb1f88b82b21f5cb9dcace33ffea3f95a189015/googleapiclient/http.py#L139
This leads to the logging output saying "rateLimitExceeded" while
HttpError's string representation contains the more detailed "Quota
exceeded for quota metric ..." - really confusing)