Move the lambda execution role into this repository
This is in preparation for adding the "lambda:InvokeFunction" permission
to the role so the Lambda itself can invoke more copies of itself to
process data concurrently.

The file terraform/modules/lambda-execution-role/lambda_iam.tf was
previously in our internal Terraform repository, where it was part of
the Concourse role previously used to deploy this service.