Update tests to reflect changes in script.py
We're essentially testing three separate invocations of the lambda now.
I moved the class out of the test function into the body to make it
easier to reference (and also changed its name to prevent aliasing (?)
the class inside test_get_group_info(), though that might not have been
necessary.